+++ decidim-surveys/spec/commands/decidim/surveys/answer_survey_spec.rb
@@ -12,6 +12,7 @@ module Decidim
       let(:survey) { create(:survey, component: component) }
       let(:survey_question_1) { create(:survey_question, survey: survey) }
       let(:survey_question_2) { create(:survey_question, survey: survey) }
+      let(:survey_question_3) { create(:survey_question, survey: survey) }
       let(:answer_options) { create_list(:survey_answer_option, 5, question: survey_question_2) }
       let(:answer_option_ids) { answer_options.pluck(:id).map(&:to_s) }
       let(:form_params) do
@@ -23,13 +24,18 @@ module Decidim
             },
             {
               "choices" => [
-                { "answer_option_id" => answer_option_ids[0], "body" => "This" },
-                { "answer_option_id" => answer_option_ids[1], "body" => "is" },
-                { "answer_option_id" => answer_option_ids[2], "body" => "my" },
-                { "answer_option_id" => answer_option_ids[3], "body" => "second" },
-                { "answer_option_id" => answer_option_ids[4], "body" => "answer" }
+                { "answer_option_id" => answer_option_ids[0], "body" => "My" },
+                { "answer_option_id" => answer_option_ids[1], "body" => "second" },
+                { "answer_option_id" => answer_option_ids[2], "body" => "answer" }
               ],
               "question_id" => survey_question_2.id
+            },
+            {
+              "choices" => [
+                { "answer_option_id" => answer_option_ids[3], "body" => "Third", "position" => 0 },
+                { "answer_option_id" => answer_option_ids[4], "body" => "answer", "position" => 1 }
+              ],
+              "question_id" => survey_question_3.id
             }
           ],
           "tos_agreement" => "1"
@@ -69,15 +75,16 @@ module Decidim
         it "creates a survey answer for each question answered" do
           expect do
             command.call
-          end.to change(SurveyAnswer, :count).by(2)
-          expect(SurveyAnswer.all.map(&:survey)).to eq([survey, survey])
+          end.to change(SurveyAnswer, :count).by(3)
+          expect(SurveyAnswer.all.map(&:survey)).to eq([survey, survey, survey])
         end
 
         it "creates answers with the correct information" do
           command.call
 
           expect(SurveyAnswer.first.body).to eq("This is my first answer")
-          expect(SurveyAnswer.last.choices.pluck(:body)).to eq(%w(This is my second answer))
+          expect(SurveyAnswer.second.choices.pluck(:body)).to eq(%w(My second answer))
+          expect(SurveyAnswer.third.choices.pluck(:body, :position)).to eq([["Third", 0], ["answer", 1]])
         end
       end
     end